Fixes #14 - Scrolldown triggered when focusing/clicking on terminal
This was a similar issue https://github.com/mui-org/material-ui/issues/16740. We focus on the hidden
input when the terminal is focused to capture input. The hidden input used to have absolute position,
which causes the browser to scroll into view. The solution was to use fixed position instead.

Also scrollIntoView was happening relative to the entire page. Passed the block: "nearest" option
into #scrollIntoView so it happens relative to the terminal div.